+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 - DefaultListSelection.DefaultItems.getValue() bug fixed, now excludes invalid indexes.
 ### is.codion.swing.common.ui
 - AbstractComponentBuilder.popupControls() bug fixed, on longer creates submenu for Controls without a caption.
+- FilterTable.Builder.columnToolTips() added.
 ### is.codion.swing.framework.ui
 - EntityConditionComponents.createEqualForeignKeyField(), combo box model only populated if it hasn't been already.